[refactor] #91 - 도메인 단위를 gitbal의 페이지로 잡았지만 종속되지 않은 친구들 global로 옮기는 작업진행
