+++ src/main/java/gitbal/backend/domain/auth/service/LoginService.java
@@ -2,17 +2,17 @@ package gitbal.backend.domain.auth.service;
 
 import gitbal.backend.domain.auth.dto.JoinRequestDto;
 import gitbal.backend.domain.auth.dto.UserDto;
-import gitbal.backend.domain.entity.Grade;
-import gitbal.backend.domain.entity.User;
+import gitbal.backend.global.entity.Grade;
+import gitbal.backend.global.entity.User;
 import gitbal.backend.domain.repository.UserRepository;
-import gitbal.backend.domain.service.MajorLanguageService;
-import gitbal.backend.domain.service.OneDayCommitService;
-import gitbal.backend.domain.service.RegionService;
-import gitbal.backend.domain.service.SchoolService;
-import gitbal.backend.domain.service.UserService;
-import gitbal.backend.exception.JoinException;
-import gitbal.backend.exception.NotDrawUserException;
-import gitbal.backend.exception.NotFoundUserException;
+import gitbal.backend.global.service.MajorLanguageService;
+import gitbal.backend.global.service.OneDayCommitService;
+import gitbal.backend.global.service.RegionService;
+import gitbal.backend.global.service.SchoolService;
+import gitbal.backend.global.service.UserService;
+import gitbal.backend.global.exception.JoinException;
+import gitbal.backend.global.exception.NotDrawUserException;
+import gitbal.backend.global.exception.NotFoundUserException;
 import gitbal.backend.global.security.CustomUserDetails;
 import gitbal.backend.global.util.AuthenticationChecker;
 import gitbal.backend.global.dto.GitbalApiDto;
