+++ src/main/java/gitbal/backend/domain/mainPage/dto/MainPageUserDto.java
@@ -1,7 +1,7 @@
 package gitbal.backend.domain.mainPage.dto;
 
 
-import gitbal.backend.domain.entity.Grade;
+import gitbal.backend.global.entity.Grade;
 
 public record MainPageUserDto(String username, long userscore, Grade grade) {
