+++ src/main/java/gitbal/backend/domain/repository/MajorLanguageRepository.java
@@ -1,7 +1,7 @@
 package gitbal.backend.domain.repository;
 
 
-import gitbal.backend.domain.entity.MajorLanguage;
+import gitbal.backend.global.entity.MajorLanguage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MajorLanguageRepository extends JpaRepository<MajorLanguage, Long> {
