+++ src/main/java/gitbal/backend/domain/repository/OneDayCommitRepository.java
@@ -1,7 +1,7 @@
 package gitbal.backend.domain.repository;
 
 
-import gitbal.backend.domain.entity.OneDayCommit;
+import gitbal.backend.global.entity.OneDayCommit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OneDayCommitRepository extends JpaRepository<OneDayCommit, Long> {
