+++ src/main/java/gitbal/backend/domain/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package gitbal.backend.domain.repository;
 
 
-import gitbal.backend.domain.entity.User;
+import gitbal.backend.global.entity.User;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
