+++ src/main/java/gitbal/backend/global/dto/PageInfoDto.java
@@ -1,4 +1,4 @@
-package gitbal.backend.domain.dto;
+package gitbal.backend.global.dto;
 
 public record PageInfoDto(int currentPage, int startPage, int endPage, int totalPage) {
