+++ src/main/java/gitbal/backend/global/entity/Grade.java
@@ -1,4 +1,4 @@
-package gitbal.backend.domain.entity;
+package gitbal.backend.global.entity;
 
 public enum Grade {
     NEWBIE, GREEN, RED, PURPLE
