+++ src/main/java/gitbal/backend/global/entity/RefreshToken.java
@@ -1,4 +1,4 @@
-package gitbal.backend.domain.entity;
+package gitbal.backend.global.entity;
 
 import lombok.AccessLevel;
 import lombok.Builder;
