+++ src/main/java/gitbal/backend/global/entity/School.java
@@ -1,4 +1,4 @@
-package gitbal.backend.domain.entity;
+package gitbal.backend.global.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,31 +6,36 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
+import java.util.Objects;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Setter //todo : MockDataGenerator에서 score값을 0으로 초기화하기 위한 세팅
-public class Region {
+public class School {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "region_id")
-  private Long regionId;
+  @Column(name = "school_id")
+  private Long id;
 
 
   @NotNull
   @Column(length = 40)
-  private String regionName;
+  private String schoolName;
 
-  @NotNull
   @ColumnDefault(value = "0")
   private Long score;
+
+
+  @ColumnDefault(value= "0")
+  private Long changedScore;
+
+  @ColumnDefault(value = "0")
+  private int schoolRank;
 
   @NotNull
   private String topContributor;
@@ -38,22 +43,20 @@ public class Region {
   @NotNull
   private Long contributorScore;
 
-
-  public Region(String regionName, Long score, String topContributor, Long contributorScore) {
-    this.regionName = regionName;
+  public School(String schoolName, Long score, String topContributor, Long contributorScore) {
+    this.schoolName = schoolName;
     this.score = score;
     this.topContributor = topContributor;
     this.contributorScore = contributorScore;
   }
 
-  public void addScore(Long score) {
-    this.score += score;
+  // TODO: test 용도여서 나중에 실제로 값 넣으면 변경해야함.
+  public static School of() {
+    return new School("광운대학교", 0L, "khyojun", 0L);
   }
 
-
-  // TODO: test 용도여서 나중에 실제로 값 넣으면 변경해야함.
-  public static Region of() {
-    return new Region("부산", 0L, "khyojun", 0L);
+  public void addScore(Long score) {
+    this.score += score;
   }
 
   public void updateContributerInfo(String nickname, Long score) {
@@ -63,5 +66,23 @@ public class Region {
     }
   }
 
+  public void setSchoolRank(int rank) {
+    this.schoolRank=rank;
+  }
 
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (!(o instanceof School school)) {
+      return false;
+    }
+    return Objects.equals(schoolName, school.schoolName);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(schoolName);
+  }
 }
