+++ src/main/java/gitbal/backend/global/entity/User.java
@@ -1,4 +1,4 @@
-package gitbal.backend.domain.entity;
+package gitbal.backend.global.entity;
 
 
 import jakarta.persistence.CascadeType;
