+++ src/main/java/gitbal/backend/global/mock/MockDataGenerator.java
@@ -1,19 +1,19 @@
 package gitbal.backend.global.mock;
 
 
-import gitbal.backend.domain.entity.Grade;
-import gitbal.backend.domain.entity.MajorLanguage;
-import gitbal.backend.domain.entity.OneDayCommit;
-import gitbal.backend.domain.entity.Region;
-import gitbal.backend.domain.entity.School;
-import gitbal.backend.domain.entity.User;
+import gitbal.backend.global.entity.Grade;
+import gitbal.backend.global.entity.MajorLanguage;
+import gitbal.backend.global.entity.OneDayCommit;
+import gitbal.backend.global.entity.Region;
+import gitbal.backend.global.entity.School;
+import gitbal.backend.global.entity.User;
 import gitbal.backend.domain.repository.MajorLanguageRepository;
 import gitbal.backend.domain.repository.OneDayCommitRepository;
 import gitbal.backend.domain.repository.RegionRepository;
 import gitbal.backend.domain.repository.SchoolRepository;
 import gitbal.backend.domain.repository.UserRepository;
-import gitbal.backend.domain.service.RegionService;
-import gitbal.backend.domain.service.SchoolService;
+import gitbal.backend.global.service.RegionService;
+import gitbal.backend.global.service.SchoolService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
